Fix Potential Alignment Issues
Add functionality for resetting database file or upload directory if one or the other is missing. This ensure a level of stability and alignment between the two facets. For example, if database file is remove and a new database is created and user attempts to add entry that is currently in the uploaded directory then the program will crash. This handles issues similar to that or vice versa if upload directory is removed and database file is not.